+++ src/provider/SemanticTokensProvider/SemanticTokensProvider.ts
@@ -2,12 +2,14 @@ import * as vscode from 'vscode';
 import type { TAhkFileData } from '../../core/ProjectManager';
 import { pm } from '../../core/ProjectManager';
 import { CMemo } from '../../tools/CMemo';
+import { ClassHighlight } from './ClassHighlight';
 import { DAList2SemanticHighlight } from './DAList2SemanticHighlight';
 import { funcHighlight } from './funcHighlight';
 import { legacyAssignmentHighlight } from './legacyAssignmentHighlight';
 import { ModuleVarSemantic } from './ModuleVarSemantic';
 import { MultilineHighlight } from './MultilineHighlight';
-import { pushToken, TokenModifiers, TokenTypes } from './tools';
+import type { TSemanticTokensLeaf } from './tools';
+import { TokenModifiers, TokenTypes } from './tools';
 // https://code.visualstudio.com/api/language-extensions/semantic-highlight-guide#standard-token-types-and-modifiers
 
 export const legend: vscode.SemanticTokensLegend = new vscode.SemanticTokensLegend(
@@ -15,21 +17,19 @@ export const legend: vscode.SemanticTokensLegend = new vscode.SemanticTokensLege
     [...TokenModifiers],
 );
 
-const Semantic = new CMemo<TAhkFileData, vscode.SemanticTokens>((AhkFileData: TAhkFileData): vscode.SemanticTokens => {
-    const { AST, ModuleVar, DocStrMap } = AhkFileData;
+const Semantic = new CMemo<TAhkFileData, readonly TSemanticTokensLeaf[]>(
+    (AhkFileData: TAhkFileData): readonly TSemanticTokensLeaf[] => {
+        const { AST, ModuleVar, DocStrMap } = AhkFileData;
 
-    const tokensBuilder: vscode.SemanticTokensBuilder = new vscode.SemanticTokensBuilder(legend);
-
-    pushToken([
-        ...DAList2SemanticHighlight(AST),
-        ...ModuleVarSemantic(ModuleVar),
-        ...funcHighlight(DocStrMap),
-        ...MultilineHighlight(DocStrMap),
-        ...legacyAssignmentHighlight(DocStrMap),
-    ], tokensBuilder);
-
-    return tokensBuilder.build();
-});
+        return [
+            ...DAList2SemanticHighlight(AST),
+            ...ModuleVarSemantic(ModuleVar),
+            ...funcHighlight(DocStrMap),
+            ...MultilineHighlight(DocStrMap),
+            ...legacyAssignmentHighlight(DocStrMap),
+        ];
+    },
+);
 
 // semantic token type
 export const AhkFullSemanticHighlight: vscode.DocumentSemanticTokensProvider = {
@@ -41,6 +41,24 @@ export const AhkFullSemanticHighlight: vscode.DocumentSemanticTokensProvider = {
         const AhkFileData: TAhkFileData | null = pm.updateDocDef(document);
         if (AhkFileData === null) return null;
 
-        return Semantic.up(AhkFileData);
+        const tokensBuilder: vscode.SemanticTokensBuilder = new vscode.SemanticTokensBuilder(legend);
+        for (
+            const { range, tokenType, tokenModifiers } of [
+                // limited memo, related to the file system
+                ...ClassHighlight(AhkFileData),
+
+                // safe memo
+                ...Semantic.up(AhkFileData),
+            ]
+        ) {
+            // Obtain first, render first, no weight api found yet
+            tokensBuilder.push(
+                range,
+                tokenType,
+                tokenModifiers,
+            );
+        }
+
+        return tokensBuilder.build();
     },
 };